% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_map.R
\name{gss_map}
\alias{gss_map}
\title{Map a given set of GSS codes}
\usage{
gss_map(
  data,
  col_gss = "gss_code",
  col_return = NULL,
  vintage = NULL,
  limit = NULL
)
}
\arguments{
\item{data}{A data frame containing a column with gss data, or a vector of
gss codes or names. These should be provided in plotting order.}

\item{col_gss}{If \code{data} is a data frame, the name of the column
containing gss codes or names.}
}
\value{
The input data frame or vector with an added column of sf objects
}
\description{
Map a given set of GSS codes
}
