% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ons_sparql_prefixes.R
\name{ons_sparql_prefixes}
\alias{ons_sparql_prefixes}
\title{Shortcuts to tables used in the ONS SPARQL queries}
\usage{
ons_sparql_prefixes()
}
\description{
Shortcuts to tables used in the ONS SPARQL queries
}
