% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_ons.R
\name{query_ons}
\alias{query_ons}
\title{Run a SPARQL query on the ONS Geography Linked Data API}
\usage{
query_ons(query, ...)
}
\arguments{
\item{query}{Character. SPARQL query for the API.}

\item{...}{Other arguments passed to SPARQL::SPARQL().}
}
\description{
The query returns a data frame with the `results` element of the list return
by the API query. For documentation of the API see
http://statistics.data.gov.uk/sparql.
}
